backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pydantic[email]
email-validator
python-jose[cryptography] 